Cast getpid() return to unsigned long to stop compilers from whining. Possibly bogus, but without a smarter configuration script we have no other option.
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@80615 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/http_main.c
@@ -339,7 +339,8 @@ static void expand_lock_fname(pool *p)
 {
     char buf[20];
 
-    ap_snprintf(buf, sizeof(buf), ".%lu", getpid());
+    /* XXXX possibly bogus cast */
+    ap_snprintf(buf, sizeof(buf), ".%lu", (unsigned long)getpid());
     lock_fname = pstrcat(p, server_root_relative(p, lock_fname), buf, NULL);
 }
 #endif
